% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_precision}
\alias{bound_precision}
\title{Bounding Numerical Precision}
\usage{
bound_precision(vals)
}
\arguments{
\item{vals}{A \code{numeric} vector of values in the interval [0, 1].}
}
\description{
Bounds extreme values to numerical (machine) precision, for use with
sensitive quantities like estimated propensity scores.
}
\keyword{internal}
