% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{cv_eif}
\alias{cv_eif}
\title{EIF for stochastic (in)direct effects with intermediate confounding}
\usage{
cv_eif(
  fold,
  data_in,
  contrast,
  g_learners,
  e_learners,
  m_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  w_names,
  m_names
)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{data_in}{A \code{data.table} containing the observed data with columns
are in the order specified by the NPSEM (Y, M, Z, A, W), with column names
set appropriately based on the input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{NULL} has
no effect, as the value of the argument \code{effect} is instead used to
define the contrasts. To override \code{effect}, provide a \code{numeric}
double vector, giving the values of a' and a* (e.g., \code{c(0, 1)}.}

\item{g_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, for use in fitting a model for the propensity score, i.e.,
\eqn{g = P(A | W)}.}

\item{e_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a cleverly parameterized propensity
score that includes the mediators, i.e., \eqn{e = P(A | Z, W)}.}

\item{m_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting the outcome regression.}

\item{q_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, for use in fitting a regression involving the
mediator-outcome confounder, i.e., \eqn{q(Z | A', W)}.}

\item{r_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., \eqn{r(Z | A', M, W)}.}

\item{u_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a reduced regression, i.e.,
\eqn{u(Z, A, W) = E[m(A, Z, M, W) * (q(Z|A,W) / r(Z|A,M,W)) *
(e(a'|M,W) / e(A|M,W)) | Z = z, A = a, W = w]}.}

\item{v_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a reduced regression, i.e.,
\eqn{v(A,W) = E[\int_m m(a', Z, M, W) * q(z|A',W) d\nu(m) | A=a, W=w)]}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by \code{\link{medoutcon}}.}
}
\description{
EIF for stochastic (in)direct effects with intermediate confounding
}
\keyword{internal}
