% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_plugin}
\alias{est_plugin}
\title{Plug-in estimator}
\usage{
est_plugin(v_pred)
}
\arguments{
\item{v_pred}{A \code{numeric} vector of the predicted values of the v(a, w)
nuisance parameter.}
}
\value{
A \code{numeric} representing the plug-in estimate of the estimand.
}
\description{
A convenience function for the plug-in estimator.
}
\keyword{internal}
