% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_nuisance_d}
\alias{fit_nuisance_d}
\title{Fit estimated efficient influence function conditioning on W, A, Z, R, Y}
\usage{
fit_nuisance_d(
  train_data,
  valid_data,
  contrast,
  learners,
  b_out,
  g_out,
  h_out,
  q_out,
  r_out,
  u_out,
  v_out,
  m_names,
  w_names
)
}
\arguments{
\item{train_data}{A \code{data.table} containing observed data, with columns
in the order specified by the NPSEM (Y, M, R, Z, A, W), with column names
set appropriately based on the input data. Such a structure is a
convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Not optional for this nuisance parameter.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{c(0, 1)}
assumes a binary intervention node \code{A}.}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}, to be used in fitting a model for this nuisance parameter.}

\item{b_out}{Output from the internal function for fitting the outcome
regression \code{\link{fit_out_mech}}.}

\item{g_out}{Output from the internal function for fitting the treatment
mechanism without conditioning on mediators \code{\link{fit_treat_mech}}.}

\item{h_out}{Output from the internal function for fitting the treatment
mechanism conditioning on the mediators \code{\link{fit_treat_mech}}.}

\item{q_out}{Output from the internal function for fitting the mechanism of
the intermediate confounder while conditioning on the mediators, i.e.,
\code{\link{fit_moc_mech}}, setting \code{type = "q"}.}

\item{r_out}{Output from the internal function for fitting the mechanism of
the intermediate confounder without conditioning on mediators, i.e.,
\code{\link{fit_moc_mech}}, setting \code{type = "r"}.}

\item{u_out}{Output from the internal function for fitting the pseudo-outcome
regression conditioning on mediator-outcome confounder , i.e.,
\code{\link{fit_nuisance_u}}.}

\item{v_out}{Output from the internal function for fitting the pseudo-outcome
regression conditioning on treatment and baseline i.e.,
\code{\link{fit_nuisance_v}}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by a call to the wrapper function \code{\link{medoutcon}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}
}
\description{
This function estimates the conditional efficient influence function,
setting the treatment to \ifelse{html}{\out{a<sup>*</sup>}}{\eqn{a^\star}}
and conditioning on W, A, Z, R, and Y. It is used to adjust the full-data
efficient influence function to account for two-phase sampling.
}
\keyword{internal}
