% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_nuisance_v}
\alias{fit_nuisance_v}
\title{Fit pseudo-outcome regression conditioning on treatment and baseline}
\usage{
fit_nuisance_v(
  train_data,
  valid_data,
  contrast,
  learners,
  b_out,
  q_out,
  m_names,
  w_names
)
}
\arguments{
\item{train_data}{A \code{data.table} containing observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Not optional for this nuisance parameter.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{c(0, 1)}
assumes a binary intervention node \code{A}.}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}, to be used in fitting a model for this nuisance parameter.}

\item{b_out}{Output from the internal function for fitting the outcome
regression \code{\link{fit_out_mech}}.}

\item{q_out}{Output from the internal function for fitting the mechanism of
the intermediate confounder while conditioning on the mediators, i.e.,
\code{\link{fit_moc_mech}}, setting \code{type = "q"}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by a call to the wrapper function \code{\link{medoutcon}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}
}
\description{
Fit pseudo-outcome regression conditioning on treatment and baseline
}
