% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_out_mech}
\alias{fit_out_mech}
\title{Fit outcome regression}
\usage{
fit_out_mech(
  train_data,
  valid_data = NULL,
  contrast,
  learners,
  m_names,
  w_names
)
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data, with
columns in the order specified by the NPSEM (Y, M, Z, A, W), with column
names set based on the original input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated \code{\link{medoutcon}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default of \code{c(0, 1)} assumes
a binary intervention node \code{A}.}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}, to be used in fitting the outcome regression, i.e., b(A,Z,M,W).}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by \code{\link{medoutcon}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}
}
\description{
Fit outcome regression
}
