% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms_moc.R
\name{fit_r_mech}
\alias{fit_r_mech}
\title{Fit mediator-outcome confounder score while conditioning on mediators}
\usage{
fit_r_mech(data, valid_data = NULL, contrast, lrnr_stack, z_names,
  w_names)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, L, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medshift}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{contrast}{...}

\item{lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a model for the
propensity score, i.e., g = P(A | W).}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}
}
\description{
Fit mediator-outcome confounder score while conditioning on mediators
}
