% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_treat_mech}
\alias{fit_treat_mech}
\title{Fit propensity scores for treatment contrasts}
\usage{
fit_treat_mech(train_data, valid_data = NULL, contrast, learners,
  w_names, m_names, type = c("g", "e"))
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data; columns
are in the order specified by the NPSEM (Y, M, Z, A, W), with column names
set appropriately based on the input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medoutcon}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{train_data}, to be used for
estimation via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention "A" to be compared. The default value of \code{c(0, 1)} assumes
binary intervention node "A", though support for categorical interventions
is planned for future releases.}

\item{learners}{\code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a propensity score model,
i.e., g = P(A = 1 | W).}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medoutcon}.}
}
\description{
Fit propensity scores for treatment contrasts
}
