% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_u_param}
\alias{fit_u_param}
\title{Fit nuisance parameter u(z,a,w)}
\usage{
fit_u_param(data, valid_data, contrast, lrnr_stack, m_output, q_output,
  r_output, e_output)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medoutcon}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{c(0, 1)}
assumes a binary intervention node \code{A}, though support for categorical
interventions is planned for future releases.}

\item{lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a model for the
propensity score, i.e., g = P(A | W).}

\item{m_output}{...}

\item{q_output}{...}

\item{r_output}{...}

\item{e_output}{...}
}
\description{
Fit nuisance parameter u(z,a,w)
}
