% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrate_over_g}
\alias{integrate_over_g}
\title{Numerical Integration for Weighted Treatment Mechanism}
\usage{
integrate_over_g(g_mech, a_vals, weighting, int_grid_prop = 0.5)
}
\arguments{
\item{g_mech}{The estimated conditional density corresponding to the natural
or shifted value of the treatment for modified treatment policies based on
the observed values of the treatment.}

\item{a_vals}{The observed values of the treatment used in computing the
conditional density estimates provided in the argument \code{g_mech}.}

\item{weighting}{A \code{numeric} vector of weights corresponding to various
regression functions estiated based on the observed values of the treatment.
Such values correspond to a multiplicative weight applied to the estimated
conditional density.}

\item{int_grid_prop}{A \code{numeric} scalar corresponding to the propotion
of points to be used in numerically evaluating an integral over the domain
of the natural/shifted conditional density of the treatment. This is only
relevant in the case of modified treatment policies for continuous-valued
exposures; it is irrelevant for incremental propensity score interventions.}
}
\description{
In the case of modified treatment policies, it is necessary to numerically
evaluate an integral over the domain of the treatment mechanism. This is a
simple procedure to numerically compute such an integral based on Monte
Carlo importance sampling from a uniform distribution.
}
\keyword{internal}
