% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_mechanisms.R
\name{ipsi_shift}
\alias{ipsi_shift}
\title{Incremental Propensity Score Intervention}
\usage{
ipsi_shift(gn_est, delta)
}
\arguments{
\item{gn_est}{...}

\item{delta}{...}
}
\description{
compute shift where gn := g(A = 1 | W)
}
