% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medoutcon.R
\name{medoutcon_likelihood}
\alias{medoutcon_likelihood}
\title{Stochastic Mediation Likelihood Factors}
\usage{
medoutcon_likelihood(tmle_task, learner_list)
}
\arguments{
\item{tmle_task}{A \code{tmle3_Task} object specifying the data and the
NPSEM for use in constructing elements of TML estimator.}

\item{learner_list}{A \code{list} specifying which learners are to be applied
for each of the regression tasks required for the TML estimator.}
}
\description{
Stochastic Mediation Likelihood Factors
}
\keyword{internal}
