% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medoutcon.R
\name{medoutcon_npsem}
\alias{medoutcon_npsem}
\title{NPSEM for Stochastic Mediation Effects with Mediator-Outcome Confounding}
\usage{
medoutcon_npsem(node_list, variable_types = NULL)
}
\arguments{
\item{node_list}{A \code{list} object specifying the different nodes in the
non-parametric structural equation model.}

\item{variable_types}{Used to define how variables are handled. Optional.}
}
\description{
NPSEM for Stochastic Mediation Effects with Mediator-Outcome Confounding
}
\keyword{internal}
