% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medshift_moc.R
\name{medshift_moc}
\alias{medshift_moc}
\title{Nonparametric estimation of decomposition term for causal mediation analysis
with stochastic interventions under mediator-outcome confounding}
\usage{
medshift_moc(W, A, L, Z, Y, contrast = c(0, 1),
  g_lrnrs = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  e_lrnrs = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  q_lrnrs = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  r_lrnrs = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  u_lrnrs = sl3::Lrnr_glm_fast$new(),
  v_lrnrs = sl3::Lrnr_glm_fast$new(), estimator = c("onestep", "tmle",
  "sub", "ipw"), estimator_args = list(cv_folds = 5))
}
\arguments{
\item{W}{A \code{matrix}, \code{data.frame}, or similar corresponding to a
set of baseline covariates.}

\item{A}{A \code{numeric} vector corresponding to a treatment variable. The
parameter of interest is defined as a location shift of this quantity.}

\item{L}{A \code{numeric} vector corresponding to a mediator-outcome
confounder affected by treatment (on the causal pathway between intervention
A, mediator Z, and outcome Y, but unaffected itself by the mediator Z).}

\item{Z}{A \code{numeric} vector, \code{matrix}, \code{data.frame}, or
similar corresponding to a set of mediators (on the causal pathway between
the intervention A and the outcome Y).}

\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{contrast}{...}

\item{g_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a model for the propensity
score, i.e., g = P(A | W).}

\item{e_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., e = P(A | Z, W).}

\item{q_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., q(L | A', W).}

\item{r_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., r(L | A', M, W).}

\item{u_lrnrs}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e., u(L, A, W) =
E[m(A, L, Z, W) * (q(L|A,W) / r(L|A,Z,W)) * (e(a'|Z,W) / e(A|Z,W)) |
L = l, A = a, W = w].}

\item{v_lrnrs}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e., v(A,W) =
E[\int_z m(a', l, Z, W) * q(l|A',W) d\nu(z) | A = a, W = w)].}

\item{estimator}{The desired estimator of the natural direct effect to be
computed. Currently, choices are limited to a substitution estimator, a
re-weighted estimator, and an efficient one-step estimator. The interested
user should consider consulting Díaz & Hejazi (2019+) for a comparative
investigation of each of these estimators.}

\item{estimator_args}{A \code{list} of extra arguments to be passed (via
\code{...}) to the function call for the specified estimator. The default
is so chosen as to allow the number of folds used in computing the AIPW
estimator to be easily tweaked. Refer to the documentation for functions
\code{\link{est_onestep}}, \code{\link{est_ipw}}, and
\code{\link{est_substitution}} for details on what other arguments may be
specified through this mechanism.}
}
\description{
Nonparametric estimation of decomposition term for causal mediation analysis
with stochastic interventions under mediator-outcome confounding
}
