% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_mechanisms.R
\name{mtp_shift}
\alias{mtp_shift}
\title{Modified Treatment Policy Shift}
\usage{
mtp_shift(A, W = NULL, delta)
}
\arguments{
\item{A}{...}

\item{W}{...}

\item{delta}{...}
}
\description{
compute shift where d(A,W) = A + delta
}
