% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_stochastic_medoutcon.R
\docType{data}
\name{Param_medshift}
\alias{Param_medshift}
\title{Parameter for mediation effect under stochastic interventions}
\format{\code{\link{R6Class}} object.}
\usage{
Param_medshift
}
\value{
\code{Param_base} object
}
\description{
Parameter definition class. See https://arxiv.org/abs/1901.02776
}
\section{Constructor}{

  \code{define_param(Param_medoutcon, shift_param, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}}
          corresponding to the observed likelihood.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be
          treated as the outcome
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood under the
          stochastic intervention on mediators and treatment.
    }
    \item{\code{lf_exptilt}}{Object derived from \code{\link{LF_base}} for
          assessing the stochastic mediation intervention.
    }
    \item{\code{treatment_task}}{\code{\link{tmle3_Task}} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{\link{tmle3_Task}} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
}
}

\concept{Parameters}
\keyword{data}
