% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halglmnet.R
\name{predict.SL.halglmnet}
\alias{predict.SL.halglmnet}
\title{predict.SL.halglmnet}
\usage{
\method{predict}{SL.halglmnet}(
  object,
  newdata,
  remove_extra_cols = FALSE,
  add_missing_cols = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted object of class \code{glmnet}.}

\item{newdata}{A matrix of new observations on which to obtain predictions.}

\item{remove_extra_cols}{A \code{logical} indicating whether extra columns
should be removed.}

\item{add_missing_cols}{A \code{logical} indicating whether missing columns
should be added to make the \code{newdata} and training data consistent.}

\item{...}{Placeholder (ignored).}
}
\value{
A \code{numeric} vector of predictions from a \code{SL.halglmnet}
 object based on the provided \code{newdata}.
}
\description{
Predict method for objects of class \code{SL.halglmnet}
}
