% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.medshift}
\alias{print.medshift}
\title{Print Method for Stochastic Mediation Parameter Objects}
\usage{
\method{print}{medshift}(x, ...)
}
\arguments{
\item{x}{An object of class \code{medshift}.}

\item{...}{Other options (not currently used).}
}
\description{
The \code{print} method for objects of class \code{medshift}.
}
