% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_original}
\alias{scale_to_original}
\title{Scale Values to the Original Scale}
\usage{
scale_to_original(scaled_vals, max_orig, min_orig)
}
\arguments{
\item{scaled_vals}{A \code{numeric} vector of values scaled to lie in the
interval [0, 1] by use of \code{\link{scale_to_unit}}.}

\item{max_orig}{The maximum of the values on the original scale.}

\item{min_orig}{The minimum of the values on the original scale.}
}
\description{
Scale Values to the Original Scale
}
\keyword{internal}
