% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_unit}
\alias{scale_to_unit}
\title{Scale Values to the Unit Interval [0, 1]}
\usage{
scale_to_unit(vals)
}
\arguments{
\item{vals}{A \code{numeric} vector of values to be scaled into the interval
[0, 1].}
}
\description{
Scale Values to the Unit Interval [0, 1]
}
\keyword{internal}
