% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halglmnet.R
\name{SL.halglmnet}
\alias{SL.halglmnet}
\title{Custom Highly Adaptive Lasso for SuperLearner}
\usage{
SL.halglmnet(
  Y,
  X,
  newX,
  family,
  obsWeights,
  id,
  alpha = 1,
  nfolds = 5,
  nlambda = 1000,
  useMin = TRUE,
  loss = "mse",
  ...
)
}
\arguments{
\item{Y}{A \code{numeric} of outcomes.}

\item{X}{A \code{matrix} of predictors/covariates.}

\item{newX}{A matrix of new observations on which to obtain predictions. The
default of \code{NULL} computes predictions on training inputs \code{X}.}

\item{family}{Not used by the function directly, but meant to ensure
compatibility with \code{SuperLearner}.}

\item{obsWeights}{Not used by the function directly, but meant to ensure
compatibility with \code{SuperLearner}. These are passed to
\code{\link[glmnet]{cv.glmnet}}.}

\item{id}{A \code{numeric} of observation IDs.}

\item{alpha}{A \code{numeric} indicating the mixing parameter for the Lasso
(L1) and Ridge (L2) penalties. The default of \code{alpha = 1} is for the
Lasso L1 penalty.}

\item{nfolds}{Integer for the number of folds to be used when splitting the
data for cross-validation. This defaults to 5.}

\item{nlambda}{A \code{numeric} corresponding to the length of the sequence
to be used for selecting an optimal value of the penalization parameter.}

\item{useMin}{Determines which lambda is selected from \code{cv.glmnet}.
\code{TRUE} corresponds to \code{"lambda.min"} and \code{FALSE} corresponds
to \code{"lambda.1se"}.}

\item{loss}{A \code{character} indicating the loss function that ought to be
considered. Choices \code{"mse"} for mean squared error, \code{"deviance"},
and any others available through \code{\link[glmnet]{cv.glmnet}}.}

\item{...}{Placeholder (ignored).}
}
\value{
An object of class \code{SL.halglmnet} with a fitted \code{glmnet}
 object and corresponding predictions based on the input data.
}
\description{
Wrapper for \pkg{SuperLearner} for a customized Highly Adaptive Lasso
}
