% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.medoutcon}
\alias{summary.medoutcon}
\title{Summary for interventional mediation effect estimate objects}
\usage{
\method{summary}{medoutcon}(object, ..., ci_level = 0.95)
}
\arguments{
\item{object}{An object of class \code{medoutcon}, as produced by invoking
\code{\link{medoutcon}}.}

\item{...}{Other arguments. Not currently used.}

\item{ci_level}{A \code{numeric} indicating the level of the confidence
interval to be computed.}
}
\description{
Print a convenient summary for objects of \code{S3} class \code{medoutcon}.
}
