% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.medshift}
\alias{summary.medshift}
\title{Summary for Stochastic Mediation Parameter Objects}
\usage{
\method{summary}{medshift}(object, ..., ci_level = 0.95)
}
\arguments{
\item{object}{An object of class \code{medshift}, as produced by invoking
the function \code{tmle_medshift}, for which a confidence interval is to be
computed.}

\item{...}{Other arguments. Not currently used.}

\item{ci_level}{A \code{numeric} indicating the level of the confidence
interval to be computed.}
}
\description{
Print a convenient summary for objects of \code{S3} class \code{medshift}.
}
