% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medoutcon.R
\docType{data}
\name{tmle3_Spec_medoutcon}
\alias{tmle3_Spec_medoutcon}
\title{Defines a TML Estimator for Outcome under Joint Stochastic Intervention on
Treatment and Mediator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_medoutcon
}
\description{
Defines a TML Estimator for Outcome under Joint Stochastic Intervention on
Treatment and Mediator
}
\keyword{datasets}
