% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{two_phase_eif}
\alias{two_phase_eif}
\title{Two-phase sampling adjusted, un-centered efficient influence function}
\usage{
two_phase_eif(R, two_phase_weights, eif, eif_predictions, plugin_est)
}
\arguments{
\item{R}{A \code{logical} vector indicating whether an sampled observation's
mediators were measured using a two-phase sampling design.}

\item{two_phase_weights}{A \code{numeric} vector of known observation-level
weights corresponding to the inverse probability of the mediators being
measured. These weights should only be provided if the two-phase sampling
indicator \code{R} is specified.}

\item{eif}{A \code{numeric} vector of the efficient influence function.}

\item{eif_predictions}{A \code{numeric} vector of the predicted efficient
influence function, conditioning on the mediator being measured.}

\item{plugin_est}{A \code{numeric} corresponding to the plug-in estimate for
the given contrast.}
}
\value{
An un-centered efficeint influence function that accounts for the
  two-phase sampling design.
}
\description{
Adjust the efficient influence function to account for the use of two-phase
sampling designs in measuring the mediators.
}
\keyword{internal}
