% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_onestep.R
\name{cv_eif}
\alias{cv_eif}
\title{Cross-validated EIF for (in)direct effect with mediator-outcome confounder}
\usage{
cv_eif(fold, data_in, lrnr_stack_g, lrnr_stack_e, lrnr_stack_m,
  lrnr_stack_phi, w_names, m_names)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{data_in}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medshift}.}

\item{lrnr_stack_g}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a model for the
propensity score, i.e., g = P(A | W).}

\item{lrnr_stack_e}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., e = P(A | Z, W).}

\item{lrnr_stack_m}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting the outcome regression,
i.e., m(A, Z, W).}

\item{lrnr_stack_phi}{A \code{Stack} object, or other learner class
(inheriting from \code{Lrnr_base}), containing a single or set of
instantiated learners from the \code{sl3} package, to be used in fitting a
reduced regression useful for computing the efficient one-step estimator,
i.e., phi(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) | W).}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}
}
\description{
Cross-validated EIF for (in)direct effect with mediator-outcome confounder
}
\keyword{internal}
