% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_onestep.R
\name{cv_eif}
\alias{cv_eif}
\title{Cross-validated EIF for (in)direct effect with mediator-outcome confounder}
\usage{
cv_eif(fold, data_in, g_lrnrs, e_lrnrs, m_lrnrs, q_lrnrs, r_lrnrs, u_lrnrs,
  v_lrnrs, w_names, m_names)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{data_in}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medshift}.}

\item{g_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a model for the propensity
score, i.e., g = P(A | W).}

\item{e_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., e = P(A | Z, W).}

\item{m_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting the outcome regression.}

\item{q_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., q(L | A', W).}

\item{r_lrnrs}{A \code{Stack} object, or other learner class (inheriting from
\code{Lrnr_base}), containing a single or set of instantiated learners from
the \code{sl3} package, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., r(L | A', M, W).}

\item{u_lrnrs}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e., u(L, A, W) =
E[m(A, L, Z, W) * (q(L|A,W) / r(L|A,Z,W)) * (e(a'|Z,W) / e(A|Z,W)) |
L = l, A = a, W = w].}

\item{v_lrnrs}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e., v(A,W) =
E[\int_z m(a', l, Z, W) * q(l|A',W) d\nu(z) | A = a, W = w)].}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}
}
\description{
Cross-validated EIF for (in)direct effect with mediator-outcome confounder
}
\keyword{internal}
