% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_onestep.R
\name{cv_eif}
\alias{cv_eif}
\title{EIF for (in)direct effect under mediator-outcome confounding}
\usage{
cv_eif(fold, data_in, contrast, g_lrnr_stack, e_lrnr_stack, m_lrnr_stack,
  q_lrnr_stack, r_lrnr_stack, u_lrnr_stack, v_lrnr_stack, w_names, m_names)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{c(0, 1)}
assumes a binary intervention node \code{A}, though support for categorical
interventions is planned for future releases.}

\item{g_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, for use in fitting a model for the propensity
score, i.e., g = P(A | W).}

\item{e_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., e = P(A | Z, W).}

\item{m_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting the outcome regression.}

\item{q_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a regression involving
the mediator-outcome confounder, i.e., q(Z | A, W).}

\item{r_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a regression involving
the mediator-outcome confounder, i.e., r(Z | A, M, W).}

\item{u_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
appearing in the efficient influence function, i.e., u(L, A, W) =
E[m(A, Z, M, W) * (q(Z|A,W) / r(Z|A,M,W)) * (e(a'|M,W) / e(A|M,W)) |
Z = z, A = a, W = w].}

\item{v_lrnr_stack}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e., v(A,W) =
E[\int_z m(a', z, M, W) * q(z | A', W) d\nu(z) | A = a, W = w)].}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}

\item{data_ing}{A \code{data.table} containing the observed data with columns
are in the order specified by the NPSEM (Y, M, Z, A, W), with column names
set appropriately based on the input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medoutcon}.}
}
\description{
EIF for (in)direct effect under mediator-outcome confounding
}
\keyword{internal}
