% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{est_onestep}
\alias{est_onestep}
\title{One-step for interventional (in)direct effects with intermediate confounding}
\usage{
est_onestep(
  data,
  contrast,
  g_learners,
  e_learners,
  m_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  w_names,
  m_names,
  y_bounds,
  ext_weights = NULL,
  cv_folds = 10
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{NULL} has
no effect, as the value of the argument \code{effect} is instead used to
define the contrasts. To override \code{effect}, provide a \code{numeric}
double vector, giving the values of a' and a* (e.g., \code{c(0, 1)}.}

\item{g_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, for use in fitting a model for the propensity score, i.e.,
\eqn{g = P(A | W)}.}

\item{e_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a cleverly parameterized propensity
score that includes the mediators, i.e., \eqn{e = P(A | Z, W)}.}

\item{m_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting the outcome regression.}

\item{q_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, for use in fitting a regression involving the
mediator-outcome confounder, i.e., \eqn{q(Z | A', W)}.}

\item{r_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a regression involving the
mediator-outcome confounder, i.e., \eqn{r(Z | A', M, W)}.}

\item{u_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a reduced regression, i.e.,
\eqn{u(Z, A, W) = E[m(A, Z, M, W) * (q(Z|A,W) / r(Z|A,M,W)) *
(e(a'|M,W) / e(A|M,W)) | Z = z, A = a, W = w]}.}

\item{v_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from \pkg{sl3}, to be used in fitting a reduced regression, i.e.,
\eqn{v(A,W) = E[\int_m m(a', Z, M, W) * q(z|A',W) d\nu(m) | A=a, W=w)]}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by \code{\link{medoutcon}}.}

\item{y_bounds}{A \code{numeric} double indicating the minimum and maximum
observed values of the outcome variable Y prior to its being re-scaled to
the unit interval.}

\item{ext_weights}{A \code{numeric} vector of observation-level weights that
have been computed externally. Such weights are used in the construction of
a re-weighted estimator.}

\item{cv_folds}{A \code{numeric} integer specifying the number of folds to
be created for cross-validation. Use of cross-validation allows for entropy
conditions on the one-step estimator to be relaxed. For compatibility with
\code{\link[origami]{make_folds}}, this value specified must be greater
than or equal to 2; the default is to create 10 folds.}
}
\description{
One-step for interventional (in)direct effects with intermediate confounding
}
