% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tml.R
\name{est_tml}
\alias{est_tml}
\title{TMLE for stochastic (in)direct effects under intermediate confounding}
\usage{
est_tml(data, contrast, g_learners, e_learners, m_learners, q_learners,
  r_learners, u_learners, v_learners, w_names, m_names,
  ext_weights = NULL, cv_folds = 10, max_iter = 5)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{\link{medoutcon}}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{NULL} has
no effect, as the value of the argument \code{effect} is instead used to
define the contrasts. To override \code{effect}, provide a \code{numeric}
double vector, giving the values of a' and a* (e.g., \code{c(0, 1)}.}

\item{g_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, for use in fitting a model for the propensity
score, i.e., \eqn{g = P(A | W)}.}

\item{e_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., \eqn{e = P(A | Z, W)}.}

\item{m_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting the outcome regression.}

\item{q_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, for use in fitting a regression involving the
mediator-outcome confounder, i.e., \eqn{q(L | A', W)}.}

\item{r_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a regression involving
the mediator-outcome confounder, i.e., \eqn{r(L | A', M, W)}.}

\item{u_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e.,
\eqn{u(L, A, W) = E[m(A, L, Z, W) * (q(L|A,W) / r(L|A,Z,W)) *
(e(a'|Z,W) / e(A|Z,W)) | L = l, A = a, W = w]}.}

\item{v_learners}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a reduced regression
useful for computing the efficient one-step estimator, i.e.,
\eqn{v(A,W) = E[\int_z m(a', l, Z, W) * q(l|A',W) d\nu(z) | A = a, W = w)]}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function
\code{\link{medoutcon}}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by a call to the wrapper function \code{\link{medoutcon}}.}

\item{ext_weights}{A \code{numeric} vector of observation-level weights that
have been computed externally. Such weights are used in the construction of
a re-weighted one-step estimator or in solving a re-weighted estimating
equation in the case of the TML estimator. Use with caution.}

\item{cv_folds}{A \code{numeric} integer value specifying the number of folds
to be created for cross-validation. Use of cross-validation / cross-fitting
allows for entropy conditions on the one-step estimator to be relaxed. Note:
for compatibility with \code{origami::make_folds}, this value specified here
must be greater than or equal to 2; the default is to create 10 folds.}

\item{max_iter}{A \code{numeric} integer giving the maximum number of steps
to be taken for the iterative procedure to construct a TML estimator.}
}
\description{
TMLE for stochastic (in)direct effects under intermediate confounding
}
