% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_tml}
\alias{est_tml}
\title{TML estimator for stochastic interventional (in)direct effects}
\usage{
est_tml(
  data,
  contrast,
  g_learners,
  h_learners,
  b_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  w_names,
  m_names,
  y_bounds,
  effect_type = c("interventional", "natural"),
  svy_weights = NULL,
  cv_folds = 5L,
  max_iter = 5L,
  tiltmod_tol = 10
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{NULL} has
no effect, as the value of the argument \code{effect} is instead used to
define the contrasts. To override \code{effect}, provide a \code{numeric}
double vector, giving the values of a' and a*, e.g., \code{c(0, 1)}.}

\item{g_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a model for the propensity score.}

\item{h_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a model for a parameterization of the
propensity score that conditions on the mediators.}

\item{b_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a model for the outcome regression.}

\item{q_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a model for a nuisance regression of
the intermediate confounder, conditioning on the treatment and potential
baseline covariates.}

\item{r_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a model for a nuisance regression of
the intermediate confounder, conditioning on the mediators, the treatment,
and potential baseline confounders.}

\item{u_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a pseudo-outcome regression required
for in the efficient influence function.}

\item{v_learners}{A \code{\link[sl3]{Stack}} object, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing instantiated
learners from \pkg{sl3}; used to fit a pseudo-outcome regression required
for in the efficient influence function.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by \code{\link{medoutcon}}.}

\item{y_bounds}{A \code{numeric} double indicating the minimum and maximum
observed values of the outcome variable Y prior to its being re-scaled to
the unit interval.}

\item{effect_type}{A \code{character} indicating whether components of the
interventional or natural (in)direct effects are to be estimated. In the
case of the natural (in)direct effects, estimation of several nuisance
parameters is unnecessary.}

\item{svy_weights}{A \code{numeric} vector of observation-level weights that
have been computed externally. Such weights are used in the construction of
a re-weighted estimator.}

\item{cv_folds}{A \code{numeric} value specifying the number of folds to be
created for cross-validation. Use of cross-validation allows for entropy
conditions on the TML estimator to be relaxed. Note: for compatibility with
\code{\link[origami]{make_folds}}, this value  must be greater than or
equal to 2; the default is to create 10 folds.}

\item{max_iter}{A \code{numeric} integer giving the maximum number of steps
to be taken for the iterative procedure to construct a TML estimator.}

\item{tiltmod_tol}{A \code{numeric} indicating the maximum step size to be
taken when performing TMLE updates based on logistic tilting models. When
the step size of a given update exceeds this value, the update is avoided.}
}
\description{
TML estimator for stochastic interventional (in)direct effects
}
