% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_nuisance_u}
\alias{fit_nuisance_u}
\title{Fit pseudo-outcome regression conditioning on mediator-outcome confounder}
\usage{
fit_nuisance_u(
  train_data,
  valid_data,
  learners,
  b_out,
  q_out,
  r_out,
  g_out,
  h_out,
  w_names
)
}
\arguments{
\item{train_data}{A \code{data.table} containing observed data, with columns
in the order specified by the NPSEM (Y, M, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medoutcon}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. NOT optional for this nuisance parameter.}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}, to be used in fitting a model for this nuisance parameter.}

\item{b_out}{Output from the internal function for fitting the outcome
regression \code{\link{fit_out_mech}}.}

\item{q_out}{Output from the internal function for fitting the mechanism of
the intermediate confounder while conditioning on mediators, i.e.,
\code{\link{fit_moc_mech}}, setting \code{type = "q"}.}

\item{r_out}{Output from the internal function for fitting the mechanism of
the intermediate confounder without conditioning on mediators, i.e.,
\code{\link{fit_moc_mech}}, setting \code{type = "r"}.}

\item{g_out}{Output from the internal function for fitting the treatment
mechanism without conditioning on mediators \code{\link{fit_treat_mech}}.}

\item{h_out}{Output from the internal function for fitting the treatment
mechanism conditioning on the mediators \code{\link{fit_treat_mech}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}
}
\description{
Fit pseudo-outcome regression conditioning on mediator-outcome confounder
}
