% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_treat_mech}
\alias{fit_treat_mech}
\title{Fit propensity scores for treatment contrasts}
\usage{
fit_treat_mech(
  train_data,
  valid_data = NULL,
  contrast,
  learners,
  m_names,
  w_names,
  type = c("g", "h")
)
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data; columns
are in the order specified by the NPSEM (Y, M, Z, A, W), with column names
set appropriately based on the input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated \code{\link{medoutcon}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{train_data}, to be used for
estimation via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{contrast}{A \code{numeric} double indicating the two values of the
intervention \code{A} to be compared. The default value of \code{c(0, 1)}
assumes a binary intervention node \code{A}.}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}, to be used in fitting a propensity score models, i.e., g :=
P(A = 1 | W) and h := P(A = 1 | M, W).}

\item{m_names}{A \code{character} vector of the names of the columns that
correspond to mediators (M). The input for this argument is automatically
generated by \code{\link{medoutcon}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medoutcon}}.}

\item{type}{A \code{character} indicating which of the treatment mechanism
variants to estimate. Option \code{"g"} corresponds to the propensity score
g(A|W) while option \code{"h"} conditions on the mediators h(A|M,W).}
}
\description{
Fit propensity scores for treatment contrasts
}
