% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{two_phase_eif}
\alias{two_phase_eif}
\title{Two-phase sampling adjusted, un-centered efficient influence function}
\usage{
two_phase_eif(S, two_phase_weights, eif, eif_predictions, plugin_est)
}
\arguments{
\item{S}{A \code{logical} vector indicating whether an observation's mediator
was measured in a two-phase sampling design. Defaults to \code{NULL},
implying that two-phase sampling did not occur.}

\item{two_phase_weights}{A \code{numeric} vector of known observation-level
weights corresponding to the inverse probability of the mediator being
measured. These weights should only be provided if \code{S} is
is specified. Defaults to \code{NULL}.}

\item{eif}{A \code{numeric} vector of the efficient influence function.}

\item{eif_predictions}{A \code{numeric} vector of the predicted efficient
influence function, conditioning on the mediator being measured.}

\item{plugin_est}{A \code{numeric} representing the plug-in estimate for the
given contrast.}
}
\value{
An un-centered
}
\description{
Two-phase sampling adjusted, un-centered efficient influence function
}
